+++ asynchronix/Cargo.toml
@@ -61,7 +61,7 @@ rmp-serde = { version = "1.1", optional = true }
 serde = { version = "1", optional = true }
 
 # gRPC service dependencies.
-tokio = { version = "1.0", features=["net"], optional = true }
+tokio = { version = "1.0", features=["net", "rt-multi-thread"], optional = true }
 tonic = { version = "0.11", default-features = false, features=["codegen", "prost"], optional = true }
 
 # WASM service dependencies.
